refactor(pipeline): refactor the artifact module of the data pipeline
- [x] move artifact configuration to a separate submodule;
